MAGE-1618 Reorganize index related service classes
