+++ Test/Unit/Service/Product/BatchQueueProcessorTest.php
@@ -13,7 +13,7 @@ use Algolia\AlgoliaSearch\Model\IndexMover;
 use Algolia\AlgoliaSearch\Model\IndicesConfigurator;
 use Algolia\AlgoliaSearch\Model\Queue;
 use Algolia\AlgoliaSearch\Service\AlgoliaCredentialsManager;
-use Algolia\AlgoliaSearch\Service\IndexSettingsComparator;
+use Algolia\AlgoliaSearch\Service\Index\Settings\IndexSettingsComparator;
 use Algolia\AlgoliaSearch\Service\Product\BatchQueueProcessor;
 use Algolia\AlgoliaSearch\Service\Product\IndexBuilder;
 use Algolia\AlgoliaSearch\Service\Product\IndexOptionsBuilder;
