+++ ViewModel/Adminhtml/IndexingManager/Section.php
@@ -3,7 +3,7 @@
 namespace Algolia\AlgoliaSearch\ViewModel\Adminhtml\IndexingManager;
 
 use Algolia\AlgoliaSearch\Helper\ConfigHelper;
-use Algolia\AlgoliaSearch\Service\IndexNameFetcher;
+use Algolia\AlgoliaSearch\Service\Index\IndexNameFetcher;
 use Algolia\AlgoliaSearch\Service\StoreNameFetcher;
 use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Framework\View\Element\Block\ArgumentInterface;
